test(ApplicationStartupActions): Fix test strategy
Previous approach was to insert a mocked root logger in the
logback setup at test runtime. However that can not work since
the sequence of initialisation goes:

  Spring Test Context -> @Before method -> @Test method

Even if we mock the root logger in the @Before method, it's too
late. The spring context has already been started and we want
to capture a message logged at the startup of the context.

Instead, i've used the spring boot logging.file property
to tell logback it should log to maven's target/ dir. The
test is then trivial, especially thanks to AssertJ's fluent
assertions which natively can handle files!

This does leave some considerations. The application's
default configuration now tries to log to a directory which
only exists at integration test time. In a future commit
i'll have separate properties for integration tests and
for regular service running. In any event this can
be overridden with an external properties file or a
command line parameter, thanks to spring boot.

The other consideration has been that i'd like to assert
the logging level since that really is an important aspect
of the feature. That's not easily possible with AssertJ
today but i may revisit this in future.